Add sawtooth module
Add a band-limited sawtooth generator. This is the same algorithm as
in music-synthesizer-for-android (that was included in the code but
never hooked up to the UI to generate sound).

The generator should probably be made more generic in the waveform,
but that's future work.

Also fixes bug in raising of biquad.