Use cpal and midir on non-Mac platforms
Use portable sound and midi libraries for platforms other than Mac.
(They would work for Mac as well, but I'm keeping that port around
because it might be useful for Audio Units later).

Also some tweaking of Send traits; it should be possible to send the
entire graph, which is necessary to run it in the cpal playback
callback.