Update copyright messages and remove CLA
Attribute copyright to The Synthesizer IO authors. This is consistent
with Google's open source guidelines, but also reflects the new status
as a personal project. The requirement for a Contributor License
Agreement is also removed.

A slight bit more text added to README, but there's lots more that will
need to be written as the project is filled out.