Make patcher widget generate deltas
Have the patcher widget generate delta events, which in turn will be
used to drive the audio engine, as opposed to only locally mutating
the model.

Right now there's a listener for the deltas, but all it does is debug
print.

Also snap to latest version of xi-win to get minor improvements.