Start on scope display
DO NOT COMMIT AS IS, Cargo.lock points to local tree

Add a scope display to the UI. Right now it just displays a sine wave
(it's not hooked up to the synthesizer engine), and doesn't update in an
animation; also the size of the computed bitmap is hard-coded, rather
than adapting to the layout size, but it's a start.

Also updates dependencies because we need direct2d 0.2 for the bitmap
drawing.