Start wiring up scope to audio
This patch doesn't wire up the audio, but does some refactoring to
get data flowing from the engine to the UI. At the moment, this is
just the count of received items.

Now that we are regularly polling, items don't leak.

The SynthState struct moves from being a listener to being a widget,
so that other listeners can synchronously access its state. Also, the
construction of the UI is factored out, so not as much is in main().